# Update qchem demos to be JAX or JAX/JIT compatible