--- a/demonstrations/tutorial_how_to_quantum_just_in_time_compile_vqe_catalyst.py
+++ b/demonstrations/tutorial_how_to_quantum_just_in_time_compile_vqe_catalyst.py
@@ -12,7 +12,7 @@ atoms sharing two electrons) using `Catalyst <https://github.com/PennyLaneAI/Cat
 quantum just-in-time framework for PennyLane, that allows hybrid quantum-classical workflows to be
 compiled, optimized, and executed with a significant performance boost.
 
-.. figure:: ../_static/demonstration_assets/how_to_vqe_qjit/OGthumbnail_large_how-to-vqe-qjit_2024-04-23.png
+.. figure:: ../_static/demo_thumbnails/opengraph_demo_thumbnails/OGthumbnail_large_how-to-vqe-qjit_2024-04-23.png
     :align: center
     :width: 60%
     :target: javascript:void(0)
@@ -44,7 +44,7 @@ We will break the implementation into three steps:
 #
 
 import pennylane as qml
-from pennylane import numpy as np
+import numpy as np
 
 dataset = qml.data.load('qchem', molname="H3+")[0]
 H, qubits = dataset.hamiltonian, len(dataset.hamiltonian.wires)