**Title:**
Update qchem demos to be JAX or JAX/JIT compatible

**Summary:**
As we phase out autograd, we need to ensure that our current qchem demos can work with either vanilla numpy or jax. Demos that are JIT compatible were made to use JIT as well. For demos that don't have any differentiation, we opt to use vanilla numpy.

Notes: the qubit tapering demo is not JIT compatible due to 2 issues: 1) Wire ordering is not ascending (Not a bug, but might need attention) 2) qml.taper_operation() uses the `Exp` operator, which has a conditional statement ([Bug report](https://github.com/PennyLaneAI/pennylane/issues/5993))

**Relevant references:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-69776]