@austingmhuang 2024-08-08
> Really nice work @austingmhuang! I have two quick questions here.
> 
> 1. Out of curiosity, how did you decide when to use vanilla numpy, use JAX without JIT, use JAX with JIT, or use qjit?

Vanilla numpy: Demos with obviously 0 differentiable requirements e.g. initial state preparation. If there was even a little potential for _maybe_ using jnp, I opted to use jnp, even if the demo actually works perfectly fine with vanilla numpy.

Jax without JIT: This was only for qubit tapering, which uses the `Exp` operator at some point in the demo (not jit compatible). If/when that bug is addressed we might revisit this demo and make it jit compatible as well.

Jax with JIT vs qjit: I used qjit as the default since that's what I saw documented in [this demo](https://pennylane.ai/qml/demos/tutorial_how_to_quantum_just_in_time_compile_vqe_catalyst/), but I'm not sure exactly what the difference between the two is to be honest. The one time I used `jax.jit`, I meant to use `qjit` 😅 .

> 2. It would be really valuable once this PR is building to have a look at the execution times for these demos with these changes, and compare them to the execution time on the live website.

👍 agree!