@josh146 2024-08-08
> Vanilla numpy: Demos with obviously 0 differentiable requirements e.g. initial state preparation. If there was even a little potential for _maybe_ using jnp, I opted to use jnp, even if the demo actually works perfectly fine with vanilla numpy.

Note that `jnp` instead of `np` often results in no change to performance (and can sometimes be slower!). So my typical rule of thumb is to only use `jnp` if I need gradients _and/or_ I want to JIT (for performance).

In most cases, you often want to import both regular numpy and JAX numpy, and use `np` for the parts of the workflow that don't need to be differentiated/dynamic, which can increase performance.

> Jax without JIT: This was only for qubit tapering, which uses the `Exp` operator at some point in the demo (not jit compatible). If/when that bug is addressed we might revisit this demo and make it jit compatible as well.

Nice. As mentioned above, `jnp` without `jit` makes sense as long as gradients are needed.

> Jax with JIT vs qjit: I used qjit as the default since that's what I saw documented in [this demo](https://pennylane.ai/qml/demos/tutorial_how_to_quantum_just_in_time_compile_vqe_catalyst/), but I'm not sure exactly what the difference between the two is to be honest. The one time I used `jax.jit`, I meant to use `qjit` 😅 .

There are a few key differences:

- `qjit` uses an entirely different execution pipeline to `jax.jit` (LLVM vs XLA). This is not user-facing, but can be important to know when developing.

- `qjit` is more performant when used with lightning, since it directly calls lightning at the C++ layer (`jax.jit` will callback into python).

- `qjit` allows you to have additional features:

  - classical control around quantum operations via `qml.for_loop`, `qml.cond`, and `qml.while_loop`.
  - `@qjit(autograph=True)` even allows _python control flow_ like `if` and `for` statements to be applied around quantum operations! See https://docs.pennylane.ai/projects/catalyst/en/stable/dev/autograph.html
  - With `@qjit` you can do arbitrary classical processing of mid-circuit measurements.
 
On the last point, being able to 'capture' this control flow around quantum operations (without unrolling for loops, for example) can make `qjit` programs much more efficiently represented, which can lead to improvements in compilation time.