@austingmhuang 2024-08-09
> Note that `jnp` instead of `np` often results in no change to performance (and can sometimes be slower!). So my typical rule of thumb is to only use `jnp` if I need gradients _and/or_ I want to JIT (for performance).

I see! Interesting, I didn't know that 🐙 

> There are a few key differences:
> 
> * `qjit` uses an entirely different execution pipeline to `jax.jit` (LLVM vs XLA). This is not user-facing, but can be important to know when developing.
> * `qjit` is more performant when used with lightning, since it directly calls lightning at the C++ layer (`jax.jit` will callback into python).
> * `qjit` allows you to have additional features:
>   
>   * classical control around quantum operations via `qml.for_loop`, `qml.cond`, and `qml.while_loop`.
>   * `@qjit(autograph=True)` even allows _python control flow_ like `if` and `for` statements to be applied around quantum operations! See https://docs.pennylane.ai/projects/catalyst/en/stable/dev/autograph.html
>   * With `@qjit` you can do arbitrary classical processing of mid-circuit measurements.
> 
> On the last point, being able to 'capture' this control flow around quantum operations (without unrolling for loops, for example) can make `qjit` programs much more efficiently represented, which can lead to improvements in compilation time.

Seems like `qjit` has a lot of great features I didn't know about! Maybe I missed it but I couldn't find a detailed guide into qjit/catalyst and its usage outside of a few demos 😅 . 

I have another question that's maybe a bit off topic but is there a difference between `jax.grad` and `catalyst.grad`?