@josh146 2024-08-12
> Seems like qjit has a lot of great features I didn't know about! Maybe I missed it but I couldn't find a detailed guide into qjit/catalyst and its usage outside of a few demos 😅 .

Best resource is likely the Catalyst docs: https://docs.pennylane.ai/projects/catalyst/en/latest/dev/sharp_bits.html

> I have another question that's maybe a bit off topic but is there a difference between jax.grad and catalyst.grad?

Yes:

- `jax.grad` is a JAX gradient transformation that directly transforms the JAXPR in Python to insert the rules for backpropagation. You can use `jax.grad` inside `@qjit`, but _only_ on classical functions (since `jax.grad` doesn't know how to differentiate quantum operations).

- `catalyst.grad` does not apply in Python; it is simply a primitive that marks within the JAXPR 'compute the gradient of this block'. The gradient is instead computed during LLVM compilation. During LLVM compilation, we have our own compiler passes for quantum gradients, and use Enzyme for classical gradients.