@josh146 2024-08-12
| Demo                                                | Time before              | Time after               |
|-----------------------------------------------------|--------------------------|--------------------------|
| Adaptive circuits                                   | 4 minutes 37.101 seconds | 4 minutes 10.302 seconds |
| Modelling chemical reactions                        | 1 minutes 4.119 seconds  | 5 minutes 55.044 seconds |
| Classically-boosted variational quantum eigensolver | 0 minutes 0.413 seconds  | 0 minutes 1.111 seconds  |
| Givens rotations for quantum chemistry              | 0 minutes 0.024 seconds  | 0 minutes 1.118 seconds  |
| Initial state preparation for quantum chemistry     | 0 minutes 9.520 seconds  | 1 minutes 7.418 seconds  |
| Measurement optimization                            | 0 minutes 12.339 seconds | 0 minutes 13.915 seconds |
| Qubit tapering                                      | 0 minutes 0.744 seconds  | 0 minutes 3.034 seconds  |
| VQE in different spin sectors                       | 0 minutes 0.567 seconds  | 0 minutes 2.283 seconds  |

It looks like:

- Modelling chemical reactions
- Initial state preparation for quantum chemistry

are worth looking into 🤔